requirements.txt
Flask==3.1.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1
scikit-surprise==1.1.4